requirements.txt
"""
requirements.txt - Python dependencies for PINN project
"""

# Core ML frameworks
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Optimization
optuna>=3.0.0
pymoo>=0.6.0
scikit-optimize>=0.9.0

# Visualization and data
matplotlib>=3.7.0
plotly>=5.0.0
pandas>=2.0.0
h5py>=3.9.0
xarray>=2023.0.0
netCDF4>=1.6.0

# Development and utilities
jupyter>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Optional: CFD tools
# pyfoam>=0.6.0  # Uncomment if using OpenFOAM Python interface
